pyproject.toml
[tool.poetry]
name = "send-tech-bot"
version = "0.1.0"
description = ""
authors = ["Replit user <replituser@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.0.1"
python-decouple = "^3.4"
youtube-dl = "^2021.3.14"
flask = "^1.1.2"
wikipedia = "^1.4.0"
google = "^3.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
